// src/models/enum/Typography.ts
export enum TextLevel {
    INLINE = "inline",
    PHARAGRAPH = 'P',
    H1 = 'h1',
    H2 = 'h2',
    H3 = 'h3',
    H4 = 'h4',
    H5 = 'h5',
    H6 = 'h6',
    // Span = 'span',
    // Div = 'div',
    // Bold = 'b',
    // Italic = 'i',
    // Underline = 'u',
    // Small = 'small',
    // Strong = 'strong',
    // Emphasis = 'em',
    // Quote = 'blockquote',
    // Preformatted = 'pre',
}